feat(framework): enable plugin loading and config
A new api has been added to the aurelia class called withPlugin which
allows you to specify a module id and a config. When you call
aurelia.start all plugins will be dynamically loaded and their install
methods will be called with the config object. install is options. If
present, it may return a promise.