feat(plugins): update to use new metadata locator function
The decorator/metadata system was updated and that necessitated changes
in the way that custom metadata locators were registered. This fix
updates the internals of the atscript support to this new api. Note
that this is a temporary change. AtScript will be going away, but we
didn’t want to remove it just yet in order to provide our customers
will time to deal with these external changes.