fix(all): create a property framework configuration object
This is a BREAKING CHANGE. The Plugins type has been removed and
replaced with FrameworkConfiguration. All configuration-style APIs from
the Aurelia object have been moved to the FrameworkConfiguration type.
Some APIs have been renamed to work better in this context. Anyone that
implemented a custom bootstrapper or a plugin, will need to update
their configuration code. It should be pretty straight forward.