chore(all): update to Babel 6 (ongoing)
There are two issues. The first is that the d.ts generator does not
export the LogManager const. The second, more serious, is that the
Babel transpiler generates code that causes a runtime error when using
es2015 re-exports.

If we have to, I can manually update the generated code to fix the
Babel error. I’d rather not so I’ve created a PR for Babel. We’ll see
if they handle it in the timely fashion.