doc(contact-manager): Version Consistency Change
While working through the tutorial I noticed that the original app.js file displayed was inconsistent with the final file state at the end of the tutorial.
Upon inspection of the source I speculated that naming the file 'ES 2015/ES NEXT' may have tampered with the view rendering method. (New to this so have not
investigated the interal). This commit creates a duplicate template with the separate name. I also noticed throughout that ES NEXT is often not rendered
when selected in the tutorial. I wasn't sure if this was just for future release purposes or if there is something wrong with the selection.
Would be happy to submit a more detailed issue for that. I also noticed some inconsistency with the way the dependency injection was written for the ES 2015 specification. I changed one error with semi-colon placement, and also changed the format to reflect the style in which the dependency injection tutorial describes it. I've taken a look through the issues and pull requests and didn't notice anything like this but if I missed something let me know (This is my first pull request)!

No breaking changes.