doc(the-aurelia-cli): add instructions about the 'copy-files' feat.
In copy we can specify files to be copied to the output directory. This is very useful when working with libraries that load files via http, which is what Bootstrap does with font files. The copy object works as a 'from':'to' setup, where 'from' is the file you want to copy, and 'to' is the destination