Add external dependency upgrade script
This addition to the documentation describes an alternative method of updating dependencies by specifying an external script that inspects the aurelia.json project, finds updatable dependencies and updates them. It's a useful alternative to listing all of the dependencies in an npm-script as the previous step describes.

It's possible that the list might need to be filtered to deps containing 'aurelia', but that could be left up to the implementer.

Unfortunately the script is quiet at the moment...I haven't piped through the events from the child process, something we could do if there was interest :).