fix(dts): add back LogManager export
This is a tricky spot because the d.ts gnerator doesn't actually generate this, so we add it by hand for each release. Then, TSDoc, which we use to generate the API docs, gets confused by this and generates bogus data, so we have to remove it before generating docs. Basically, two separate tools both error in different ways depending on whether it's there or not.